improving p2p robustness
This is a set of changes to improve peer to peer functionality.
Connecting to a peer first attempts to resolve the hostname. If a
connection attempt is made to an ip address and port that is already
connected, the connection attempt is aborted. Bad data received from a
peer no longer crashes the daemon, instead a warning is logged.