generate nodekey and persist to disk
This allows for creation of an ECDSA key pair (node key) for
establishing node identity and signing messages. A node key is generated
randomly and automatically if none exists. Generated node keys are saved
to disk in the xu data folder and read on future starts. Methods to
encrypt and decrypt stored keys on disk are included but currently are
not used.