add and modify npm scripts
This adds an `npm stop` script, skips compiling each time `npm start` is
called, and splits the `dev` script into `dev` - which compiles and
starts - and `dev:watch` - which watches for changes to the code while
running. This also fixes the lint script by combining two `tslint`
commands, one according to the tsconfig.json project file and one
according to specified folders with a modified config to avoid warnings
for required type information.