organize scripts, add proto script, fix compile
The `proto` script can be used to regenerate protobuf files. The compile
script has been modified to copy over generated `.js` files from the
`lib` folder to `dist`. This also fixes a bug in windows with precompile
that causes the script to fail if the `dist` folder does not already
exist. Finally, scripts are sorted alphabetically.