separate api service logic from grpc layer
This creates a separate Service class which implements all exposed API
methods and which itself is exposed directly on the Xud server instance,
thereby facilitating direct testing of API methods. All loaded proto
files use the `convertFieldsToCamelCase` option to match the camelCase
variable naming style we use throughout the project, which differs from
the recommended snake_case style for grpc fields. The GrpService class
is now simply a wrapper for the Service class using a generic unaryCall
method to trigger the callback and provide a central location for error
handling (currently extremely basic).