replace hostId with peerId on orders
This replaces the `hostId` identifier on orders with `peerId`. Since orders are no longer being persisted between sessions, and since all orders for a peer are removed once that peer is disconnected, we no longer need to associate orders with a persistent identifier like `hostId`. This will make it easier to associate matched, swapped, or invalidated orders with an actual live peer.