fix socket error & close handling
This changes the `socket` property on the `Peer` class to be nullable and adds null checks before we attempt to write to the socket. It renames the `destroy` Peer instance method to `close` to differentiate from the `Socket.destroy` method and to align with the `open` method name. It removes the `close` call from the `error` event handler, since it's called in the `close` event. As per the Node.js Net module documentation. the `close` event is called right after the `error` event. Lastly, this renames the `connect` and `disconnect` methods in Pool.ts to `open` and `close` to match the corresponding method names in Peer.ts.