fix bug causing host uniqueness violation
This fixes a bug that was causing xud to attempt to add existing hosts to the database. The `HOSTS` packet was serializing `Host` objects to JSON. Upon receiving this packet, xud would create JSON objects which did not inherit the function definitions of the `Host` class and, critically, of the `SocketAddress` class which was used as a unique identifier for hosts. The `toString()` method of `SocketAddress` was therefore lost and reverted to that of `Object`. `getOrCreateHost` would use the `toString` method to see if a Host exists, but then create using the actual `address` and `port` fields which would still exist on the object.

This commit changes the `HOSTS` packet to pass `HostFactory` typed objects, from which a new `SocketAddress` is constructed. `getOrCreateHost` asserts that `socketAddress` is an instance of `SocketAddress` to catch similar bugs from resurfacing.