initialize pool after orderbook
This moves the initialization of the p2p pool until after the orderbook is initialized. This is because peer connections will trigger `ORDERS` and `GET_ORDERS` packets that depend on an initialized orderbook to work properly. It also ensures that the p2p pool is initialized before starting the gRPC server.