remove extraneous p2p logging & closes (#301)
This removes calls of the `Peer.close` method before a socket is open, which prevents the `close` event from firing, instead relying on the `cleanup` method to tie up loose ends when a connection to a peer errors or times out. It also removes duplicate log messages when a socket errors or is closed, and prevents logging stack traces for non-fatal peer errors.