LND support for BTC and LTC (prep for atomic swap)
This PR add support for 2 LNDs (one for BTC chain and one for LTC chain).
In additon, it fixes issues related to reconnect to LND after LND restart while XUD is active. Startup swquance is not important any more.
Also, there is different handling now if case LND is not configured (not suppose to be connected ) and LND is disconnected