replace init task with seed values
This commit makes xud add seed values for pairs and currencies when a new database is created during startup. These are in addition to the seed nodes. An `initDb` config option can be set to `false` to disable initializing new databases with these seed values. This makes the `db:init` task obsolete.