refactor grpc
This uses our static generated files for creating the xud grpc server, which requires that we pass the generated gRPC classes for the callback.

The structure laid out here is that Service.ts deals with our internal types and plain javascript objects but doesn't touch anything grpc-related. It passes values to the GrpcService.ts class which in turn creates the gRPC response object and passes it to the callback - or if it encounters an error it passes the error to the callback. The logic to map internal errors to grpc errors is located within the getServiceResponse method.