configure package for distribution
This prepares the xud package for distribution and installation via `npm install -g xud` with just the compiled files needed for execution and no source code for development. The `google-protobuf` package is included in the `grpc_tools_node_protoc_ts` dev dependency, and is added as a regular dependency for distribution.