update tslint-config-airbnb
This includes some linting updates that will be helpful for when we switch to typescript 3.x, and also adds two missing rules from the base airbnb guide, one of which was explicitly defined in our tslint.json and is now removed because it's inherited. Linting unused variables is still disabled for now but a TODO indicates that it should be reenabled once development is relatively stable.