refactor swaps into new module
Creates a module to handle swap related logic and execution and moves all swap-specific logic out of Pool and Service into a class that extends EventEmitter to notify listeners when swaps fail or succeed. Also adds the ability to reject deals if the currency is not supported, such as when lnd is disabled.