relax HelloPacket checks
This allows peers to accept a `HelloPacket` and update handshake state data at any time, and not only when we are explicitly waiting for one. This addresses an edge case whereby we receive a `HelloPacket` before we've started waiting for one, resulting in a "[peer] sent an unsolicited Hello packet" error message. Previously, there was a small disconnect between when a socket was established and when waiting for the HelloPacket began. This changes it so that the wait for the HelloPacket begins immediately after the socket becoming connected and before any packets can be received. It also adds a `handshake` event that can be used going forward for Peer listeners that are interested in changes to the handshake state.