chore: generate sample config on new version
This commit updates our package.json to generate a sample config file
when bumping the version using the `npm version` script and commit it
via git. It modifies the sample config file to not include any file
paths, but adds comments to explain the default paths and directories
based on the platform. It also adds steps to run lint and test scripts
to ensure tagged versions don't contain any build errors.