refactor: remove maxResults from getOrders (#492)
This commit removes `maxResults` and `limit` from the `getOrders` rpc
call and all related internal calls. The reason for this is because
orders are not ordered according to price internally - limiting the
number of orders returned simply picks an arbitrary set of orders
rather than the orders with the best price. It is up to the client to
sort and discard orders after they've been retrieved from `xud`.