docs: add licenses to protos and update readme
This commit applies the MIT license to our proto definitions and copies
the license from the lnd project into the lndrpc.proto file. This is
because projects that incorporate the `xud` rpc API via the proto files
may be unwilling to follow the copyleft restrictions of the AGPL-3.0
license - the MIT license makes it easier for closed source and
propietary software (as many exchanges use) to incorporate our proto
files or any files generated from it.

It also updates the readme to explicitly state that the AGPL-3.0 license
applies to all files in the repository except for those with licenses
provided in the header of the file.