feat: update handshake on lnd pubkey change
This commit creates an event for when a LightningClient verifies (or
re-verifies) a connection to an lnd server. It includes the lnd
identity pub key if it changed. When the pub key changes, the new value
is broadcast to all currently connected peers within a new Hello packet.
This allows for switches of lnd servers without requiring a restart of
xud, and also will inform peers of lnd pub keys when lnd comes back
online should lnd be offline at the time of peer connection.

Closes #513.