feat(p2p): store last address used for conn (#496)
This commit stores the last address to be successfully used for an
outgoing connection, even if it is not advertsed. This address will be
the first address to be used when automatically reconnecting to a known
peer as well as the address used for the backoff connection retry logic.

Closes #365.

BREAKING CHANGE: Adds a column to the `nodes` table.