fix: don't send order invalidation to taker
This commit fixes a bug where a a maker sends an invalidation packet to
the taker after an order is executed. The taker will have already
adjusted the orderbook locally upon successful execution, resulting in
a duplicate attempt. This fix makes it so makers don't send invalidation
packets to takers.

Fixes #506. Fixes #478.