refactor: use positive quantity for buys & sells
This change refactors how we represent buy and sell orders. Previously,
sell orders were represented with a negative quantity. However, this
was considered to be potentially confusing especially in regards to
the API and quantity updates to existing orders. Instead, orders  on
the rpc layer now have a `side` property to indicate whether they are
buys or sells, and all standing orders and order matches are represented
with positive quantities. Internally, orders use an `isBuy` property.
to indicate their side.