fix(p2p): prevent new conns after closing pool
This commit ensures that no connections that are in a pending state
when the p2p pool is disconnected are allowed to complete. It closes
all pending connections and any connections that open after
disconnection of the pool. Previously, a connection that's attempted
right before disconnecting could have interfered with shutting down
`xud`.

Fixes #516.