feat(lndclient): check if chain is out of sync
This commits checks to make sure lnd is synced to the backend chain when
verifying a connection to it. If out of sync, we set a timer to reverify
the same as if we could not call `getInfo` and hit a `Disconnected`
status. This also removes the check against verifying the connection
after a connection has been verified, as we might want to make checks
going forward that lnd is still responsive and in sync.