feat: leave partially swapped orders in order book
This commit makes it so we only remove part of our own maker orders
when only part of them get swapped. Previously, the full order was
removed from the order book even if only part of it was swapped. Now,
a partial swap results in a quantity reduction but leaves the order
in the order book.