feat: check if order is swappable before broadcast
This commit adds a check before broadcasting an order to peers to
ensure that we have active swap clients for both sides of the trading
pair for that order. In other words, we check to make sure that an
order can be swapped before advertising it to peers. This is done to
avoid polluting peers' order books with orders that can't possibly be
swapped, which can result in unnecessary swap failures and banning.