refactor: treat failed swaps as invalidations (#577)
This commit changes the event emitted for failed swaps during matching
from `peerOrder.filled` to `peerOrder.invalidated`. It also doesn't add
back peer orders to the book if they could not be swapped, and it makes
the `verifyExecution` method private. This is so that we don't confuse
orders that are removed from the order book because they were
successfully swapped with those that were removed without being swapped.