fix(p2p): don't remove peer orders without pubkey (#586)
This fixes a bug where we were attempting to remove orders for a peer
when the peer disconnects pre-handshake, the outcome meant removing
own orders by accident. This adds a check to skip the peer order removal
logic when the peer does not have a pub key associated with it.

Fixes #539.