fix(p2p): consistent checks before connecting
This commit fixes the logic that prevents us from connecting to banned
nodes or creating duplicate connections to any node. Some of that logic
was only applied for connecting to a set of nodes, but not on attempts
to connect to a single node like through the rpc `connect call. All
general checks have been moved to the `addOutbound` method which is a
key method that all outgoing peer connection attempts go through.