fix(lnd): automatically retry if chain out of sync
This commit properly schedules automatic retries to verify a connection
to lnd when the backing chain is out of sync. It also adds a separate
hardcoded delay for reconnecting due to chain being out of sync, it is
longer since syncing the chain typically takes longer to resolve than
other intermittent connection issues.

Fixes #593.