refactor(orderbook): rename & restructure types
This commit restructures the order types to reuse & inherit properties
from each other wherever possible to maintain a consistent definitions
and comments/annotations. It renames the `StampedOrder` which is
used throughout the codebase to the more brief `Order` - the previous
`Order` type which was defined and used only with the orders types
file is now called `LimitOrder` to differentiate itself from
`MarketOrder`. It also creates an `IncomingOrder` type for orders that
are received from peers along with the peers' node pub keys, this
in turn is used for the `packet.order` events rather than the
`PeerOrder`/`StampedPeerOrder` type which was incorrect.

Resolves the first part of #437.