test(db): add basic db happy path tests
This introduces a test suite for the database. Currently the the tests
only go through a "happy path" of creating a handful of database objects
and ensuring that there are no foreign key constraint violations, data
type mismatches, or other errors that could arise from errors in how
the Sequelize models are set up. Going forward this test suite can be
used to verify constraints and validate that data is written to the
database as expected.