feat(orderbook/swaps): release order hold
This releases the hold placed on an order when we accept a swap once
the swap either completes or fails. Previously, holds would remain on
orders post-swap which would make part or all of the order unavailable
for future swaps.

New unit tests are added to cover adding and removing holds.

Closes #549.