feat(swaps): calculate swap amounts locally
This makes each node involved in a swap determine the swap amounts and
currencies (`makerCurrency`, `takerCurrency`, `makerAmount`, and
`takerAmount`) independently. If there is any discrepancy in these
values, the swap will fail.These values are no longer sent as part of
the `SwapRequestPacket`. Previously, only the taker (who initiates the
swap) would determine these values and the maker would blindly accept
them.

It also updates the Swap unit tests to align with the new methods
and logic used in the `beginSwap` and `acceptDeal` routines.

Closes #615.