feat(rpc): add UNAVAILABLE grpc error status code
This adds the `UNAVAILABLE` code as one of the possible grpc status
codes that can be returned by `xud`. This status is intended for calls
that can fail due to transient conditions and may be successful if tried
again at a later time. The two existing errors that meet that
description are `LND_IS_UNAVAILABLE` (renamed from
`LND_IS_DISCONNECTED` because lnd may be reachable via grpc but out of
sync with the chain)) and `COULD_NOT_CONNECT` from the p2p context.
We automatically attempt to reestablish a connection to lnd, and nodes
that are offline may come back online.