perf(orderbook): streamline matching routine
This aims to improve the efficiency and organization of the matching
logic.

Previously, when an order was partially matched, it would be split into
two new orders (and objects) and the original order would be
dereferenced. Orders in the queue that were partially matched would be
removed from the queue, split, and then added back. Adding and removing
orders executes in `O(log n)`, as opposed to `O(1)` for peeking the top
of the queue.

Instead, this handles partial matches by creating only one new object
with the matching quantity while reducing the quantity of the existing
order (similar to how order invalidations or removes from the API are
handles). Orders are only removed from the queue when they are fully
matched. By reducing the number of adds/removes to the queue and new
objects being created, this should improve performance of the matching
logic somewhat.