feat(db): set null price for market orders
This sets the price for market orders as db null when they get saved
to the database. We use prices the max number value and 0 for buy and
sell orders internally, respectively. This can be helpful for matching
purposes but it does not need to be stored in the databases.