fix(lnd): don't default lnd clients as disabled
This fixes a bug introduced in #672 whereby the lnd clients are not
properly initializing. That PR removed synchronous calls to methods that
had asynchronous alternatives, which necessitated moving code from the
synchronous `LndClient` constructor to an asynchronous init method.

Included in the code that was moved was logic to determine whether to
change the client's status from the default value of `Disabled`. This
impacted the xud initialization procedure which would only attempt
to initialize (or previously verify a connection for) an lnd client that
was not disabled.

This PR creates a new `NotInitialized` status as the default to reflect
the status of a newly created client. It gets changed when the `init`
method is called.