refactor(p2p): remove PeerList class
This removes the `PeerList` class, which was being used essentially as
a wrapper for a `Map` and was unnecessary code. Instead, `Pool` uses
a `Map` of pub keys to peers, same as `pendingOutgoingConnections`. At
one point, `PeerList` contained custom logic but it hasn't for some
time.

This also stands to be a minor performance gain as we were inefficiently
reimplementing the `Map.size` functionality with a `length` getter.