feat(rpc/orderbook): handle remove order with hold
This handles the possibility that we attempt to remove an order via the
rpc layer while that order has a hold for a pending swap. Instead of
removing the entire order immediately, this only removes the available
portion of the order and then waits until any swaps finish before
removing the portions that were on hold.

It also adds a field to the `RemoveOrder` rpc response which indicates
whether any portion of the order was on hold. This gives feedback up
front as to whether an order was removed immediately, or whether there
are ongoing swaps which are being allowed to finish.

Closes #552.