fix(proto): use 64 bit chan_id ints as strings
This adds `jstype = JS_STRING` annotations on the channel ids in the
lnd proto definition so that they use `string` types instead of numbers.
Javascript does not natively support 64 bit numbers and precision
can be lost, and the channel ids are known to cause issues because of
this.

This also adds a new package to generate type definitions from the
`.proto` files. This is because of a known bug in the previously used
plugin, agreatfool/grpc_tools_node_protoc_ts#10, which does not update
the generated type definition when `jstype = JS_STRING` is used.

Fixes #745.