fix(swaps): release order hold on sendpayment err
This fixes a bug where the hold on an order would not be released when
the swap failed after we had issued the `sendPaymentSync` command to
lnd. Doing so changes the `SwapPhase` from `SwapAgreed` to `AmountSent`
and the latter phase was overlooked in the logic that released the hold
on swap failure.

This also renames the `AmountSent` phase to `SendingAmount` to more
accurately describe the current state. Comments have also been added
to the swap phases to better describe each one. `SendingAmount` means
we've requested `lnd` to make the payment but we don't know if it
has succeeded yet.