feat(swaps): comprehensive SwapFailureReason
This adds `SwapFailureReason` values to cover every reason that a swap
may fail. Previously we only set a string error message on swaps when
they fail/error. The failure reason can be used to take different
actions according to the reason, and is a clean way to categorize the
reasons that a swap has failed.

Closes #671.