feat(rpc): remove peerpubkey from executeswap cli
This call removes the `peer_pub_key` field from the `ExecuteSwap` cli
command and makes it optional on the rpc call. It is not truly required
information and it is inconvenient to have to provide from the cli.
Finding which peer an order belongs to is not much of a performance
concern.  It is constant time in regards to the number of orders and
linear in regards to the number of peers.

Closes #694.