fix(p2p): check before emitting open event
This addresses an edge case whereby a socket is destroyed by a peer
while we are in the process of opening a peer connection and performing
the handshake. Emitting the `open` event triggers sending several
packets, which results in an `[ERR_STREAM_DESTROYED]` error and crashes
`xud`. To prevent this, we check that the socket exists and is not
destroyed before trying to send any packets.