refactor(p2p): use every() to validate packets
This uses the `every` Array method for arrays rather than `filter` to
check that every object in a list within a packet passes validation.
They function similarly but `every` does not create a new array, whereas
`filter` does, so this cuts out an unnecessary step and may provide a
very minor performance improvement.