feat(p2p/orderbook): drop orders for dropped pairs
This adds logic to drop a peer's orders for a given trading pair when
we receive an updated handshake state that does not include a trading
pair that was supported according to the previous handshake. Without
this, we may keep orders in our order book for a peer that no longer
supports (and can swap) those orders.

Closes #599.