fix(p2p): hash order invalidations correctly
This fixes a bug where the hash calculated for outgoing order
invalidation packets wouldn't match the calculated hash for the packet
received by the peer due to the order objects being hashed having
extraneous properties that are not transmitted to the peer. This would
cause the order invalidations to be ignored by the receiver.

Fixes #767.