refactor(swaps): rename errorReason swap field
This renames the `errorReason` swap deal property to `errorMessage` to
help differentiate it from `failureReason` and describe it as a field
for details about why a swap failed or what caused an error. It also
makes it optional on the `SwapFailed` packet to accommodate cases
where there are no relevant details to include besides the
`failureReason`.

BREAKING CHANGE: Renames `errorReason` field in database.