feat(p2p/db): add failureReason field for swaps
This adds the `failureReason` field to the `SwapDeal` database model
as well as the `SwapFailed` packet. It removes the `PeerFailedSwap`
failure reason, and assigns swaps the failure reason specified by the
peer in cases when the peer fails the swap. It adds a `PaymentHashReuse`
reason which was missing to cover all current swap failure reasons.

BREAKING CHANGE: New `failureReason` p2p packet and database field.