feat(swaps): timeout stalled swaps
This adds timeouts to all swap deals to fail them after a certain
amount of time. Current time limits are set to 10 seconds for a swap to
be accepted and 30 seconds for a swap to be completed once accepted.
Previously, swaps that stalled out due to an unresponsive peer or swap
client would last indefinitely and the corresponding order would remain
on hold.

Closes #653.