fix(orderbook/db): upsert orders
This uses the `upsert` command instead of `create` for the
`addOrderIfNotExists` function to prevent a uniqueness constraint error
if there's a race condition between two simultaneous calls to that
method. It is possible that both calls query and see that the order
doesn't exist in the db at the same time, and then both try to create
the same record resulting in the uniqueness constraint error.