feat(swaps): define behavior for swap errors
This commit adds logic to handle swap failures according to the reason
for the failure. Certain failures suggest a persistent, underlying issue
with swaps for that peer and trading pair, and in this case we drop
all orders for that peer and trading pair to avoid subsequent failures.
We also penalize the peer for swaps that fail due to potential
misbehavior or due to a timeout. Timeouts result in delays that degrade
the trading experience moreso than regular swap failures.

Closes #661.