refactor(orderbook): throw error in removeOwnOrder
This changes the `removeOwnOrder` method to pass on any exceptions it
encounters rather than returning `true` or `false`. In all instances
where the method is being used, the return value is ignored anyway
and the expectation is that the order removal will succeed, so
throwing exceptions when order removal fails will ensure errors are
not unintentionally ignored.