fix(p2p): toggle opened after handshake
This moves the logic for toggling the `opened` property of a peer to
occur after the handshake is complete and verified. We rely on the
`opened` property to determine whether we accept non-handshake packets,
previously it would have been possible to accidentally accept all
packets from a peer while the handshake was still ongoing.