fix(packets): correct order packet validation
This is a minor fix to the logic to validate an order packet, primarily
checking that the `order.id` field is not empty. It also removes a
duplicate check on the `id` header field and aligns the logic with the
`OrdersPacket` validation which checks that `price` and `quantity` are
not just non-zero but positive.