perf(p2p): use async randomBytes for packet framer
This replaces the synchronous usage of `crypto.randomBytes` with the
asynchronous version when framing packets. The synchronous version can
block the thread until enough entropy is collected. Since this is done
every time an encrypted packet is sent, the delay may become noticeable.

This required rewriting synchronous code as asynchronous in several
methods and tests which makes up the bulk of the code changes.