fix(p2p): peer error logging
This fixes the way certain errors are logged when attempting to connect
to peers. Previously it was possible to see messages like:

`[P2P] warn: could not open connection to peer
(xud2.test.exchangeunion.com:8885): undefined`

This would occur when the `err` object in the catch block is a `string`.
This change ensures we log `string` errors correctly.