fix(p2p): handle packets received during handshake
This commit addresses an edge case bug whereby one peer finishes the
handshake before the other and begins sending regular packets. It is
possible for the other peer to attempt to process these regular packets
while it is executing asynchronous code to complete the handshake, which
results in the regular packets being discarded as "unsolicited."

This makes two changes:

1. Moves the logic to asynchronously send a `GetNodesPacket` until after
we mark the peer as `opened`.
2. Add an `initializing` flag that is set to true once we receive a
`SessionAckPacket`. This is an indicator that any subsequent packets we
receive should not be acted upon until the handshake is complete. If
we receive packets while `opened` is false but `initializing` is true,
we wait to see which of the `open` or `close` events fire first to
either accept or discard a packet, respectively.

Fixes #839.