fix(proto): set price as double
This fixes a bug in the way the `SwapSuccess` proto message was defined
where `price` was set as an integer which is inconsistent with how
`price` is defined for other messages and with how it is represented
and being set internally. This caused the grpc response for the
`PlaceOrder` and `ExecuteSwap` calls to be corrupt and fail due to an
assertion error.