test(p2p): add connect test if we are banned
This adds a test case that isolates a bug that was fixed in #845 where
we would successfully connect to a peer that had banned us. Although
that peer would immediately disconnect with us after establishing the
connection, this was not evident to callers of the `Service.connect`
method including the gRPC API. This test fails prior to the recent fix.

Closes #735.