fix(p2p): prevent writing to closed socket
This prevents an edge case bug where the TCP socket to a peer is
destroyed as we are about to send data to that peer. This ensures that
we check not only that a socket connection exists but that it also
hasn't been destroyed. It also adds a try/catch block as a safety
measure, since an unhandled error risks crashing the process.