refactor(orderbook): throw error on non-swap fail
This tweaks the logic around handling exceptions that arise from
swap execution to ensure that the exception is a valid
`SwapFailureReason` and to clarify the log messages and intent of the
code. If the exception is not a `SwapFailureReason` we abort the
matching procedure instead of repeating it for the failed amount.