chore: use existing protoc/xudrpc.proto for sim
This modifies the gen_protos.sh script for the simulation tests to use
the `protoc` executable that exists within the `node_modules` folder and
which we already use to generate static protobuf files for xud. It also
uses the existing xudrpc.proto file so that we don't need to maintain
duplicate copies of the file within the repository.