test: fix placeOrderAndBroadcast race condition
This attempts to fix a potential race condition that occasionally caused
the simulation tests to fail. In `placeOrderAndBroadcast` we call
`SubscribeAddedOrders` on node A, then immediately place an order on
node B, which immediately broadcasts the order to node A. It's possible
that the order subscription is not active on node A by the time it hears
about the order since there's no latency between nodes.

This fix calls `ListPeers` after `SubscribeAddedOrders` to ensure that
the two nodes are still connected and moves the `getOrdersCount` call
to after `SubscribeAddedOrders`. This in turn creates a slight delay
between subscribing for added orders and adding the order, resolving
the race.

Similar changes are also made to the `SubscribeRemovedOrders` calls.