feat: add sanity swap checks after handshake
This is a major commit that changes the procedure for accepting orders
from peers to ensure that we can send and receive payments for both
currencies that make up the order's trading pair. This introduces the
concept of "sanity swaps," which are atomic swaps of 1 smallest unit for
1 smallest unit within a single currency/chain. The flow for atomic
swaps is as follows:

1. A connection is established with a peer.
2. The handshake is performed successfully.
3. A `SanitySwapPacket` will be sent to the peer to inform them of a
sanity swap attempt for each currency the appears in the peer's trading
pairs advertised in the `nodeState`.
4. We send a 1 satoshi payment to the peer for each sanity swap.
5. Upon receiving our 1 satoshi payment, the peer attempts to learn the
preimage by sending a 1 satoshi payment to us.
6. We release the preimage that we created, effectively settling the
swap.
7. After all sanity swaps resolve, we check which advertised pairs had
successful swaps for both the base and quote currency.
8. These pairs are marked "active" and we send a `GetOrdersPacket` for
the newly activated pairs.

Other notable changes include:

- Tracking "active" pairs on a per-peer basis. "Active" pairs are those
that have been advertised by a peer's node state, which we have been
able to successfully perform a sanity swap for, and which have not
encountered unexpected payment errors from our perspective.
- Not requesting all orders via a `GetOrdersPacket` automatically upon
a successful connection and handshake.
- Ignoring order packets we receive for trading pairs that are not
active.
- The `ListPeers` rpc call only returns active trading pairs.
- Sanity swaps can be disabled via the `nosanitychecks` config option.

Closes #676.

BREAKING CHANGE: New SanitySwapPacket.