refactor(rpc): combine added & removed order calls
This combines the `SubscribeAddedOrders` and `SubscribeRemovedOrders`
rpc calls into a single `SubscribeOrders` calls. This is done to ensure
that updates to the order book do not get transmitted out of order.
gRPC guarantees message ordering within an individual call:
https://grpc.io/docs/guides/concepts.html#server-streaming-rpc. However,
with separate calls an edge case exists where a client receives an order
removal message before the order added message for that same order.

Closes #903.