refactor: node file system API
This commit uses the `fs.promises` API that was introduced in Node.js
v10.0.0 in place of the versions we created using `util.promisify`.
It also replaces usage of the `exists` method where it is used to check
whether a file exists before reading or writing to it. Node.js best
practice is to attempt to read/write directly and handle errors due to
a file existing or not existing. This is because the asynchronous nature
of these calls means that the result of `exists` could be stale by the
time we attempt subsequent reads or writes.