feat(lnd): use lnd 0.6.1 w/ hold invoices
This is a major commit that changes the way hashes are resolved and
swaps are performed with lnd. Here we switch from using a fork of lnd
to the main lnd repository.

Previously, we used the hash resolver service which made a resolve gRPC
call from lnd to xud when lnd accepted an htlc with a payment hash it
did not recognize. Now, when we prepare for a swap xud will tell lnd
about the hash via `AddHoldInvoice` and subscribe to updates to that
invoice via `SubscribeSingleInvoice`. The subscription will alert xud
when lnd accepts an htlc for that invoice, at which point xud will
attempt to resolve the hash by completing the swap and provide the
preimage to lnd via `SettleInvoice`. If a swap fails, the invoice in
lnd is canceled via `CancelInvoice`.

A new SanitySwapAck packet is necessary for this new approach.
Previously, the sanity swap payment and the SanitySwap packet (which
informs the peer of the hash for the swap would) be sent at the same
time. A node that received a resolve request for an amount of 1 satoshi
and a hash it did not recognize would wait a short while to see if it
received a SanitySwap packet for that hash. With hold invoices, lnd
must be informed of the payment hash via `AddInvoice` before it receives
the htlc, otherwise it will fail the htlc immediately. Thus, the
SanitySwapAck packet is used to inform the peer that it is ready to
accept the sanity swap payment.

Closes #798.