chore: reenable v12 current node build
This reenables the current node build in our travis configuration,
reversing #902, and updates packages that did not previously support
Node.js v12.

This also includes the results of running `npm audit fix` (reverting
`grpc-tools` to 1.6.0 because of a regression bug in newer versions)
and `npm dedupe` after updating the necessary packages.