fix: switch swap capacity check currency
This fixes a bug where the outbound capacity sanity check was comparing
against the inbound currency and amount rather than outbound. In
addition, this commit refactors & renames the methods used to calculate
the amounts and currencies involved in a swap to minimize potential
confusion. It also adds tests to detect a regression in the calculation
logic.

Fixes #932.