test: cache lnd & dependencies in travis
This modifies our Travis CI configuration to cache the go folder that
holds lnd and its dependencies, such that we won't need to install
lnd for each build and will reduce test run time significantly. This
also moves the `go` folder out of the `temp` folder since this folder
need not change unless the version of lnd we're using changes, and in
that case the tests will reinstall lnd automatically. Even locally,
keeping the go folder in between test runs will save time.

The install and simulation test scripts are modified to use the go
folder created for the simulation tests as the only GOPATH value.
This means all dependency modules and btcd are put in this folder
and separated from the local go environment, preventing the
possibility for conflicts and inconsistencies due to the environment.