fix(swaps): don't retry verified currency sanity
This modifies the logic for determining which currencies to perform
sanity swaps for. Previously, if a sanity swap failed for either
currency in a trading pair, when that trading pair was retried we would
retry sanity swaps for both currencies in the pair including the one
that worked previously. This change tracks currencies for which we
have had a successful sanity swap for and prevents retrying sanity
swaps for that currency - until we encounter an error that deactivates
a trading pair that includes that currency.

Fixes #946.