fix: swap client initialization
This fixes a bug where the swap clients would fail on their first
attempt to initialize because they would try to update `pool.nodeState`
which is undefined. `pool.nodeState` is not assigned a value until
`pool.init()` is called, which would not happen until after the first
attempt to initialize swap clients.

This changes the initialization flow so that the node state is assigned
a value in the `Pool` constructor. The properties of `Pool.nodeState`
are then updated as the other components (swap clients, order book)
initialize.