fix: calculate swap amount for market orders
This fixes a bug where the calculated swap amount for a market buy order could
be infinity. This could cause the sanity capacity check to fail when checking
a market order. Market buy orders internally have a positive infinity price
value for matching logic purposes, but the final price for market orders is
not known upon first placing them.

Fixes #978.