refactor(swapclientmanager): remove pool
This removes the `pool` dependency from `SwapClientManager`, allowing
it to be instantiated before `Pool` is instantiated in the `Xud`
initialization flow. Since `pool` relies on `nodeKey`, this change paves
the way for using `LndClient` and its `GenSeed` call to create a new
`nodeKey`.

This also makes the `Swaps` module listen directly to
`SwapClientManager` for `htlcAccepted` events, rather than binding to
the swap clients it contains. This allows the `Swaps` class to be
instantiated even if the `SwapClientManager` it contains has not been
initialized yet.